Rename app.stats to app.record
