+++ src/app/record/common.rs
@@ -1,0 +1,1 @@
+pub const RECORDER_FILE_SUFFIX: &'static str = ".rc.aget";
