--- src/app/stats/mod.rs
@@ -1,2 +1,0 @@
-pub mod list_stats;
-pub mod range_stats;
